+++ packages/integration/test/testcase/digestion/Connected_Region-Parent-Disconnected_Region-Child_BROKEN.spec.ts
@@ -1,6 +1,7 @@
 import { Component } from "@cydran/cydran";
 import { Harness } from "@cydran/testsupport";
 import { describe, expect, test } from '@jest/globals';
+import LoggingSegmentDigester from "./LoggingSegmentDigester";
 
 const PARENT_TEMPLATE: string = `<div>
 	<c-region name="child1" value="m().values"></c-region>
@@ -62,13 +63,11 @@ class ChildComponent2 extends Component {
 describe("Connected Region -> Parent -> Disconnected Region -> Child", () => {
 
 	test.skip("Connected Region -> Parent -> Connected Region -> Child", () => {
-		const segmentDigester: any = null; //LoggingSegmentDigester = new LoggingSegmentDigester();
-
-		const harness: Harness<ParentComponent> = new Harness<ParentComponent>(() => new ParentComponent(), {
-			"cydran.internal.factory.segment-digester": () => segmentDigester
-		});
-
+		const harness: Harness<ParentComponent> = new Harness<ParentComponent>(() => new ParentComponent());
+		harness.registerSingletonGlobally("cydranSegmentDigester", LoggingSegmentDigester);
 		harness.start();
+
+		const segmentDigester: LoggingSegmentDigester = harness.getContext().getObject("cydranSegmentDigester");
 
 		harness.getComponent().$c().regions().set("child1", new ChildComponent1());
 		harness.getComponent().$c().regions().set("child2", new ChildComponent2());
